project/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
qrcode[pil]==7.4.2
Werkzeug==2.3.7
cryptography==41.0.4
PyMySQL==1.1.0
Pillow==10.0.0